+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.27] - 2022-1-3
+- Fix padding issues on behavior tracker
+- Fix token logger issues
 
 ## [0.18.26] - 2021-1-2
 - Fix app crashes after completing an token logger activity